+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added publicDir config
+
 ## v0.1.1 (2022-11-30)
 * Removed Systemic dependency